+++ src/main/shards/Hello-world-shard/index.ts
@@ -1,9 +1,9 @@
-import { shardFn } from '@shared/yuekui-shard/interface'
+import { BaiYueKuiShard, shardFn } from '@shared/yuekui-shard/interface'
 import { Shard } from '@shared/yuekui-shard/decorators'
 
 const SHARD_ID = 'hello-world'
 @Shard(SHARD_ID)
-export class HelloWorldShards {
+export class HelloWorldShard implements BaiYueKuiShard {
   static id = SHARD_ID
   id = SHARD_ID